task_distribution_system/requirements.txt
#django_extensions==3.1.5
djangorestframework==3.13.1
#kombu==4.6.11
Django==4.1.7
psycopg2-binary
requests==2.27.1
jsonfield==3.1.0
#django-redis==5.2.0
#marshmallow==3.14.0
#py-eureka-client==0.11.4
#Django==3.2.13
#django-redis==5.2.0
#kafka-python==2.0.2
#marshmallow==3.14.0
#uWSGI==2.0.21
